Remove legacy (v1 out-of-process) task scanner
tintoy/msbuild-project-tools-server#42
