+++ src/LanguageServer.Engine/Handlers/DocumentSyncHandler.cs
@@ -200,7 +200,7 @@ namespace MSBuildProjectTools.LanguageServer.Handlers
                     if (Workspace.Configuration.Language.CompletionsFromProject.Contains(CompletionSource.Task))
                     {
                         Log.Verbose("Scanning task definitions for project {ProjectName}...", projectDocument.ProjectFile.Name);
-                        List<MSBuildTaskAssemblyMetadata> taskAssemblies = await projectDocument.GetMSBuildProjectTaskAssemblies();
+                        List<MSBuildTaskAssemblyMetadata> taskAssemblies = projectDocument.GetMSBuildProjectTaskAssemblies();
                         Log.Verbose("Scan complete for task definitions of project {ProjectName} ({AssemblyCount} assemblies scanned).", projectDocument.ProjectFile.Name, taskAssemblies.Count);
 
                         Log.Verbose("===========================");
