+++ test/LanguageServer.Engine.Tests/MSBuildTests.cs
@@ -83,7 +83,7 @@ namespace MSBuildProjectTools.LanguageServer.Tests
         public void GetReferencedPackageVersion(string packageId, string expectedPackageVersion)
         {
             var projectFile = new FileInfo(
-                Path.Combine(TestDirectory.FullName, "..", "..", "..", "LanguageServer.Engine.Tests.csproj")
+                Path.Combine(TestDirectory.FullName, "..", "..", "..", "..", "LanguageServer.Engine.Tests.csproj")
             );
             Assert.True(projectFile.Exists,
                 $"Cannot find project file {projectFile.FullName}"
